requirements.txt
streamlit
pandas
pyyaml
faster_whisper
openai
requests
polars
yamllint